Refactor InventoryDashboard and AddProduct components: removed unused state and functions, integrated ActivityLog component, and simplified form schema in AddProduct.
